dali dataset moved to dataset interface
